feat: unified all double/float to float

+++ pendulum/src/states/calibration/as5600_state.cpp
@@ -7,9 +7,9 @@
 #include "src/utils/as5600.h"
 #include "src/utils/log_macros.h"
 
-constexpr double STABILITY_DURATION_US = 5'000'000.0;
-constexpr double LOW_PASS_ALPHA = 0.5;
-constexpr double VELOCITY_THRESHOLD = 0.4; // rad/s
+constexpr float STABILITY_DURATION_US = 5'000'000.0;
+constexpr float LOW_PASS_ALPHA = 0.5;
+constexpr float VELOCITY_THRESHOLD = 0.4; // rad/s
 
 SequenceStatus as5600_calibration(AS5600CalibrationResult *result) {
 
@@ -23,13 +23,13 @@ SequenceStatus as5600_calibration(AS5600CalibrationResult *result) {
 
 	last_sample_time = t;
 
-	static double angle_rad = 0;
-	static double angle_rad_prev = 0;
-	static double cos_angle = 0, cos_angle_prev = 0;
-	static double sin_angle = 0, sin_angle_prev = 0;
-	static double angular_vel_filtered = 0;
+	static float angle_rad = 0;
+	static float angle_rad_prev = 0;
+	static float cos_angle = 0, cos_angle_prev = 0;
+	static float sin_angle = 0, sin_angle_prev = 0;
+	static float angular_vel_filtered = 0;
 
-	static double avg_raw_angle = 0;
+	static float avg_raw_angle = 0;
 	static int avg_sample_count = 0;
 
 	static unsigned long stable_since_us = 0;
@@ -48,8 +48,8 @@ SequenceStatus as5600_calibration(AS5600CalibrationResult *result) {
 	cos_angle = cos(angle_rad);
 	sin_angle = sin(angle_rad);
 
-	double dt_s = dt / 1'000'000.0;
-	double angular_vel = ((sin_angle - sin_angle_prev) * cos_angle - (cos_angle - cos_angle_prev) * sin_angle) / dt_s;
+	float dt_s = dt / 1'000'000.0;
+	float angular_vel = ((sin_angle - sin_angle_prev) * cos_angle - (cos_angle - cos_angle_prev) * sin_angle) / dt_s;
 
 	LOOP_LOG("[CALIBRATION] [AS5600] raw_angle : %i, angular_vel: %.2f, cos: %.2f, sin: %.2f", raw_angle, angular_vel,
 	         cos_angle, sin_angle);
